desalt: phase separation
